python/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
faker==19.3.0
urllib3<2.0.0
six
